+++ source/core/src/main/com/csse3200/game/services/ServiceLocator.java
@@ -28,6 +28,7 @@ public class ServiceLocator {
   private static TimeService timeService;
   private static GameTime timeSource;
   private static GameArea gameArea;
+  private static LightService lightService;
   private static MissionManager missions;
 
   public static GameArea getGameArea() {
@@ -58,6 +59,9 @@ public class ServiceLocator {
   }
   public static TimeService getTimeService() {
     return timeService;
+  }
+  public static LightService getLightService() {
+    return lightService;
   }
 
   public static MissionManager getMissionManager() {
@@ -107,6 +111,11 @@ public class ServiceLocator {
   public static void registerMissionManager(MissionManager source) {
     logger.debug("Registering mission manager {}", source);
     missions = source;
+  }
+
+  public static void registerLightService(LightService source) {
+    logger.debug("Registering light service {}", source);
+    lightService = source;
   }
 
   public static void clear() {
